1191 Fix UnsupportedOperationException during startup
