+++ src/main/java/ai/elimu/web/servlet/CustomDispatcherServlet.java
@@ -4,17 +4,14 @@ import ai.elimu.model.v2.enums.Environment;
 import ai.elimu.model.v2.enums.Language;
 import ai.elimu.util.ConfigHelper;
 import ai.elimu.util.db.DbContentImportHelper;
+import ai.elimu.web.ConnectionProviderWeb;
 import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
 
 import ai.elimu.util.db.DbMigrationHelper;
 import ai.elimu.web.context.EnvironmentContextLoaderListener;
 import java.io.File;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
 import java.util.EnumSet;
 import javax.persistence.Entity;
 import org.apache.logging.log4j.LogManager;
@@ -70,7 +67,7 @@ public class CustomDispatcherServlet extends DispatcherServlet {
 //                .configure("META-INF/jpa-persistence.xml")
                 .applySetting("hibernate.dialect", ConfigHelper.getProperty("jpa.databasePlatform"))
                 .applySetting("hibernate.hbm2ddl.auto", "update")
-                .applySetting(AvailableSettings.CONNECTION_PROVIDER, getConnectionProvider())
+                .applySetting(AvailableSettings.CONNECTION_PROVIDER, new ConnectionProviderWeb())
                 .build();
 
         MetadataSources metadataSources = (MetadataSources) new MetadataSources(serviceRegistry);
@@ -103,33 +100,6 @@ public class CustomDispatcherServlet extends DispatcherServlet {
         schemaExport.setDelimiter(";");
         schemaExport.setFormat(true);
         schemaExport.create(EnumSet.of(TargetType.SCRIPT), metadata);
-    }
-
-    private ConnectionProvider getConnectionProvider() {
-        return new ConnectionProvider() {
-            @Override
-            public Connection getConnection() throws SQLException {
-                return DriverManager.getConnection(ConfigHelper.getProperty("jdbc.url"), ConfigHelper.getProperty("jdbc.username"), ConfigHelper.getProperty("jdbc.password"));
-            }
-
-            @Override
-            public void closeConnection(Connection connection) throws SQLException {}
-
-            @Override
-            public boolean supportsAggressiveRelease() {
-                return true;
-            }
-
-            @Override
-            public boolean isUnwrappableAs(Class aClass) {
-                return false;
-            }
-
-            @Override
-            public <T> T unwrap(Class<T> aClass) {
-                return null;
-            }
-        };
     }
 
 }
